autocomplete: move models into geocoder module
